backend/requirements.txt
django
 djangorestframework
 django-cors-headers
 gunicorn
 pillow
 django-cloudinary-storage
 cloudinary
 psycopg2-binary
 python-dotenv
 anthropic
